script/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
openai
dotenv
selenium
undetected_chromedriver
google-generativeai
pandas